simplify: PR #122 follow-up — extract firstNonEmptyLine to format.ts
The PR added a copy of firstNonEmptyLine to agent-network.ts that was
already byte-identical with the one in agents.ts. Both view files live
in packages/api/src/views/; the existing format.ts in the same dir
hosts sibling helpers (deriveShortId, formatRelativeShort). Natural
home, single source of truth.

- Add firstNonEmptyLine to packages/api/src/views/format.ts.
- Drop the local copies from agent-network.ts and agents.ts.
- Both files import it alongside their other format.ts helpers.

Skipped (with reasons):
- Tag_line JOIN SQL fragment dedup — four sites with different WHERE/
  ORDER/aggregation; cargo-cult constant would hurt readability.
- rowToAgentDisplay shared base — divergence is pre-existing and not
  worsened by this PR; bigger refactor.
- <Specialization> component — 2 callsites with different layouts
  (line-clamp-2 vs line-clamp-3, size-gated render); parametrizing
  would create more than it saves.
- BLOCK_NAMES.tag_line constant — pre-existing pattern of inline
  literals across many files; separate refactor.
